ai-backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.27.0

# AI / Embeddings
google-generativeai>=0.7.0
# If you prefer OpenAI, swap with: openai>=1.40.0

# Vector store + RAG
langchain>=0.3.0
langchain-google-genai>=1.0.0
langchain-community>=0.3.0
faiss-cpu>=1.8.0

# Database (Supabase)
supabase>=2.5.0
